// types/api/models/Reddit.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Reddit = {
  readonly id?: number
  readonly name?: string
  readonly text?: string
  readonly image?: string
  readonly url?: string
  readonly username?: string
  readonly username_url?: string
  readonly created?: string
}